+++ Modules/Dashboard/Private/Write-DashboardObject.ps1
@@ -1,4 +1,4 @@
-function Write-Object {
+function Write-DashboardObject {
 
 <#
 .SYNOPSIS